requirements.txt
pymysql
petl
pandas
psycopg2-binary
sqlalchemy